tests/wpt/web-platform-tests/css/selectors/i18n/css3-selectors-lang-054.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>[lang="cs-CZ"], lang="cs-Latn-CZ"</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-selectors/#attribute-selectors'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
.test div { width: 50px; }
#box[lang='cs-CZ'] { width: 100px; }
</style>
</head>
<body>



<div class="test"><div id="box" lang="cs-Latn-CZ">&#xA0;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, "A lang= value with language and region subtags will NOT match a lang attribute value with language, script and region subtags.");
</script>

<div id='log'></div>

</body>
</html>
